dep: remove serde/serde_json — hand-written JsonValue + clippy fixes
Final commit in the dependency removal plan:

- serde/serde_json removed from root crate (replaced by hand-written
  JsonValue enum in src/json_value.rs with recursive-descent parser,
  to_string_pretty/to_string_compact, and From conversions)
- All #[derive(Serialize, Deserialize)] and #[serde(...)] attributes
  stripped from root crate types
- LSP JSON-RPC, telemetry, diagnostics, LLM API, and test assertions
  all use the hand-written JsonValue
- Auto-fixed remaining clippy warnings (redundant closures, div_ceil,
  for-loop, derive, cast, unnecessary parens)

Final state:
- cargo tree --depth 1: ZERO external dependencies (only 10 internal
  workspace crates)
- All wave48 + wave50 tests pass

Dependency removal summary (all 9 crates eliminated):
  log       — vuma_log! macro (c5a29012)
  tempfile  — OpenOptions::create_new (86cbb941)
  chrono    — hand-written epoch→ISO8601 (320c6460)
  rayon     — std::thread::scope (445a6097)
  libc      — local C type aliases (be53c527)
  clap      — hand-written argv parser (fb7f266e)
  toml      — hand-written toml_lite parser (89ef8919)
  proptest  — deterministic property tests (eccf306f)
  serde     — hand-written JsonValue (b755eef8 + this commit)

+++ src/time.rs
@@ -43,7 +43,7 @@ const fn civil_from_days(days: i64) -> (i64, u32, u32) {
     // the month-length arithmetic.
     let z = days + 719_468;
     let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
-    let doe = (z - era * 146_097) as i64; // [0, 146_096]
+    let doe = z - era * 146_097; // [0, 146_096]
     let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146_096) / 365; // [0, 399]
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100); // [0, 365]
